# packages/contracts/gen/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "platform aggregate API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "summary": "rotate the refresh cookie (reuse kills the family) and mint a new access token",
        "responses": {
          "200": {
            "description": "rotated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "accessToken": {
                              "type": "string",
                              "maxLength": 4096
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "missing/invalid/replayed refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        },
        "x-rate-limit-class": "strict"
      }
    },
...
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "operationId": "beginMFA",
        "summary": "create a time-limited TOTP enrollment and QR code",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "responses": {
          "200": {
            "description": "enrollment started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "x-request-body-limit": 4096,
        "operationId": "verifyMFA",
        "summary": "verify TOTP and enable MFA for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$",
                    "maxLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "listSessions",
        "summary": "active sessions of the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Session"
                              }
                            },
                            "meta": {
                              "$ref": "#/components/schemas/EnvelopeMeta"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "revokeAllSessions",
        "summary": "revoke every other session of the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sessions/{id}": {
      "delete": {
        "operationId": "revokeSession",
        "summary": "revoke one session of the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "unknown session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot": {
      "post": {
        "x-request-body-limit": 4096,
        "operationId": "forgotPassword",
        "summary": "always-200 password reset request (anti-enumeration)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "uniform response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        },
        "x-rate-limit-class": "strict"
      }
    },
    "/api/v1/auth/reset": {
      "post": {
        "x-request-body-limit": 16384,
        "operationId": "resetPassword",
        "summary": "consume a single-use reset token, set new password, wipe all sessions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "password updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "400": {
            "description": "invalid/expired/consumed token or weak password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        },
        "x-rate-limit-class": "strict"
      }
    },
    "/api/v1/auth/reset/validate": {
      "post": {
        "x-request-body-limit": 4096,
        "operationId": "validateResetToken",
        "summary": "validate a reset token without consuming its single-use grant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetTokenInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "token valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "400": {
            "description": "invalid, expired, or consumed token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        },
        "x-rate-limit-class": "strict"
      }
    },
    "/api/v1/auth/users/{id}/password": {
      "post": {
        "x-request-body-limit": 16384,
        "operationId": "adminSetUserPassword",
        "summary": "admin sets a new password for a user (revokes their sessions)",
        "x-required-permission": "user:update:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "newPassword"
                ],
                "properties": {
                  "newPassword": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 72
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "password updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/confirm-password": {
      "post": {
        "x-request-body-limit": 4096,
        "operationId": "confirmPassword",
        "summary": "re-authenticate the current user before a sensitive admin action",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 72
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "401": {
            "description": "invalid password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "operationId": "changePassword",
        "summary": "change the authenticated user's password after verifying the old password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-auth": "required",
        "x-rate-limit-class": "strict",
        "x-request-body-limit": 1024,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "oldPassword",
                  "newPassword"
                ],
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 72
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 72
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "changed and all sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "401": {
            "description": "old password invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/{id}/sessions": {
      "get": {
        "operationId": "adminListUserSessions",
        "summary": "list active sessions for a managed user",
        "x-required-permission": "user:read:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "adminRevokeUserSessions",
        "summary": "force logout a managed user on every device",
        "x-required-permission": "user:update:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/{id}/sessions/{sessionId}": {
      "delete": {
        "operationId": "adminRevokeUserSession",
        "summary": "revoke one active session for a managed user",
        "x-required-permission": "user:update:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "unknown session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/{id}/state": {
      "patch": {
        "x-request-body-limit": 4096,
        "operationId": "adminSetUserState",
        "summary": "activate/deactivate and lock/unlock a managed user",
        "x-required-permission": "user:update:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  },
                  "locked": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EnvelopeMeta": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total"
        ],
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EnvelopeOK": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 255
          },
          "data": {
            "description": "endpoint-specific payload"
          }
        }
      },
      "EnvelopeFail": {
        "type": "object",
        "required": [
          "success",
          "message",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 255
          },
          "error": {
            "type": "string",
            "maxLength": 120
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userAgent": {
            "type": "string",
            "maxLength": 512
          },
          "ip": {
            "type": "string",
            "maxLength": 64
          },
          "deviceId": {
            "type": "string",
            "maxLength": 128
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "current": {
            "type": "boolean"
          }
        }
      },
...
      "ForgotInput": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        }
      },
      "ResetInput": {
        "type": "object",
        "required": [
          "token",
          "newPassword"
        ],
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 4096
          },
          "newPassword": {
            "type": "string",
            "minLength": 12,
            "maxLength": 72
          }
        }
      },
      "ResetTokenInput": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 4096
          }
        }
      },
...
    }
  }
}